Migrated logic from the router to the controller

+++ src/user/user.controller.ts
@@ -1,6 +1,7 @@
 import { UserService } from './user.service';
 import { IUser } from './user.interface';
 import { UserValidator } from './user.validator';
+import { userModel } from './user.model';
 /*
 * UserService handles the logic of the requests
 * before calling the database
@@ -10,30 +11,42 @@ export class UserController {
     if (await UserValidator.idExists(id)) {
       return await UserService.getById(id);
     }
-    throw new Error('User id does not exist');
+    throw new Error('User does not exist');
   }
-  // static getById(id: string) {
-  //   return UserService.getById(id);
-  // }
 
   static getByName(name: String) {
     return UserService.getByName(name);
   }
 
-  static update(id: String, partialUser: Partial<IUser>) {
-    return UserService.update(id, partialUser);
+  static async update(id: string, partialUser: Partial<IUser>) {
+    if (partialUser._id) {
+      if (id !== partialUser._id) {
+        throw new Error('User bad id');
+      }
+    }
+    if (await UserValidator.idExists(id)) {
+      return UserService.update(partialUser._id, partialUser);
+    }
+    throw new Error('User already exists');
   }
 
   static getAll() {
     return UserService.getAll();
   }
 
-  static add(newUser: IUser) {
-    return UserService.add(newUser);
+  static async add(reqUser) {
+    const newUser: IUser = new userModel(reqUser);
+    if (!(await UserValidator.idExists(newUser._id))) {
+      return await UserService.add(newUser);
+    }
+    throw new Error(`User already exists`);
   }
 
-  static deleteById(id: string) {
-    return UserService.deleteById(id);
+  static async deleteById(id: string) {
+    if (await UserValidator.idExists(id)) {
+      return UserService.deleteById(id);
+    }
+    throw new Error(`User does not exist`);
   }
 
   static deleteAll() {
